# docs/ROLES_API_CONTRACTS.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CASA roles governance contracts",
    "version": "1"
  },
  "paths": {
    "/api/admin/preview/lessons/{lesson_id}": {
      "get": {
        "tags": [
          "admin-preview"
        ],
        "summary": "Preview Lesson",
        "operationId": "preview_lesson_api_admin_preview_lessons__lesson_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLessonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/preview/lessons/{lesson_id}/document": {
      "get": {
        "tags": [
          "admin-preview"
        ],
        "summary": "Preview Document",
        "operationId": "preview_document_api_admin_preview_lessons__lesson_id__document_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/preview/quizzes/{quiz_id}": {
      "get": {
        "tags": [
          "admin-preview"
        ],
        "summary": "Preview Quiz",
        "operationId": "preview_quiz_api_admin_preview_quizzes__quiz_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Quiz Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuizOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/me/scopes": {
      "get": {
        "tags": [
          "admin-scopes"
        ],
        "summary": "My Scopes",
        "operationId": "my_scopes_api_admin_me_scopes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopesOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}/scopes": {
      "get": {
        "tags": [
          "admin-scopes"
        ],
        "summary": "User Scopes",
        "operationId": "user_scopes_api_admin_users__user_id__scopes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin-scopes"
        ],
        "summary": "Assign Scopes",
        "operationId": "assign_scopes_api_admin_users__user_id__scopes_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeReplacement"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminLessonDepthLevelIn": {
        "properties": {
          "depth_key": {
            "$ref": "#/components/schemas/LessonDepthKey"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "depth_key",
          "label",
          "title",
          "body"
        ],
        "title": "AdminLessonDepthLevelIn"
      },
      "AdminLessonOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Id"
          },
          "demo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Objectives",
            "default": []
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/AdminLessonSectionIn"
            },
            "type": "array",
            "title": "Sections",
            "default": []
          },
          "depth_levels": {
            "items": {
              "$ref": "#/components/schemas/AdminLessonDepthLevelIn"
            },
            "type": "array",
            "title": "Depth Levels",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "title",
          "position",
          "status"
        ],
        "title": "AdminLessonOut"
      },
      "AdminLessonSectionIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "image_alt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Alt"
          },
          "diagram": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagram"
          }
        },
        "type": "object",
        "required": [
          "title",
          "body"
        ],
        "title": "AdminLessonSectionIn"
      },
      "AdminQuestionOptionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "option_text": {
            "type": "string",
            "title": "Option Text"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "option_text",
          "is_correct"
        ],
        "title": "AdminQuestionOptionOut"
      },
      "AdminQuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/AdminQuestionOptionOut"
            },
            "type": "array",
            "title": "Options",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_text",
          "difficulty"
        ],
        "title": "AdminQuestionOut"
      },
      "AdminQuizOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "$ref": "#/components/schemas/QuizKind"
          },
          "lesson_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Id"
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Id"
          },
          "pass_threshold": {
            "type": "integer",
            "title": "Pass Threshold"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/AdminQuestionOut"
            },
            "type": "array",
            "title": "Questions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "kind",
          "pass_threshold",
          "status"
        ],
        "title": "AdminQuizOut"
      },
...
      "ContentStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "PUBLISHED",
          "ARCHIVED"
        ],
        "title": "ContentStatus"
      },
...
      "DocumentBlockOut": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "items": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "text",
          "confidence"
        ],
        "title": "DocumentBlockOut"
      },
      "DocumentSectionOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/DocumentBlockOut"
            },
            "type": "array",
            "title": "Blocks",
            "default": []
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/DocumentSectionOut"
            },
            "type": "array",
            "title": "Children",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "level",
          "confidence"
        ],
        "title": "DocumentSectionOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LessonDepthKey": {
        "type": "string",
        "enum": [
          "ESSENTIAL",
          "TECHNICAL",
          "MATHEMATICS",
          "IMPLEMENTATION",
          "ARCHITECTURE",
          "GOVERNANCE"
        ],
        "title": "LessonDepthKey"
      },
      "LessonDocumentOut": {
        "properties": {
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/DocumentSectionOut"
            },
            "type": "array",
            "title": "Sections",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "source_file",
          "page_count"
        ],
        "title": "LessonDocumentOut"
      },
...
      "QuizKind": {
        "type": "string",
        "enum": [
          "PRACTICE",
          "VALIDATION",
          "FINAL"
        ],
        "title": "QuizKind"
      },
...
      "ScopeGrantOut": {
        "properties": {
          "school_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Id"
          },
          "pathway_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pathway Id"
          },
          "assigned_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned By"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          }
        },
        "type": "object",
        "required": [
          "school_id",
          "pathway_id",
          "assigned_by",
          "assigned_at"
        ],
        "title": "ScopeGrantOut"
      },
      "ScopeReplacement": {
        "properties": {
          "school_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "School Ids"
          },
          "pathway_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Pathway Ids"
          }
        },
        "type": "object",
        "title": "ScopeReplacement"
      },
      "ScopesOut": {
        "properties": {
          "school_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "School Ids"
          },
          "pathway_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Pathway Ids"
          },
          "grants": {
            "items": {
              "$ref": "#/components/schemas/ScopeGrantOut"
            },
            "type": "array",
            "title": "Grants"
          },
          "global_access": {
            "type": "boolean",
            "title": "Global Access"
          }
        },
        "type": "object",
        "required": [
          "grants",
          "global_access"
        ],
        "title": "ScopesOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}